# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Beneficiaries API",
    "description": "Backend contract for the Arabic Beneficiaries mobile application. Authenticated endpoints use an HTTP Bearer access token. Business-controlled values such as points, attendance QR values, submission status, and duplicate awards are calculated by the server.",
    "contact": { "name": "Beneficiaries application team" },
    "license": { "name": "Proprietary" },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/events": {
      "get": {
        "tags": ["Home and Content"],
        "summary": "List upcoming published events",
        "operationId": "list_events_api_v1_events_get",
        "security": [{ "HTTPBearer": [] }],
        "parameters": [
          {
            "name": "fromDate",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                { "type": "string", "format": "date" },
                { "type": "null" }
              ],
              "title": "Fromdate"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BeneficiaryEventResponse"
                  },
                  "title": "Response List Events Api V1 Events Get"
                }
              }
            }
          },
          "400": {
            "description": "Business rule or invalid request",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "401": {
            "description": "Authentication required or invalid",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "403": {
            "description": "Authenticated but not permitted",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "404": {
            "description": "Requested resource not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "409": {
            "description": "Resource state conflict",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "422": {
            "description": "Field validation failed",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "429": {
            "description": "Rate or attempt limit exceeded",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "503": {
            "description": "External delivery service unavailable",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          }
        }
      }
    },
    "/api/v1/readings/today": {
      "get": {
        "tags": ["Home and Content"],
        "summary": "Get today's daily reading, or the latest published reading",
        "operationId": "get_daily_reading_api_v1_readings_today_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    { "$ref": "#/components/schemas/DailyReadingResponse" },
                    { "type": "null" }
                  ],
                  "title": "Response Get Daily Reading Api V1 Readings Today Get"
                }
              }
            }
          },
          "400": {
            "description": "Business rule or invalid request",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "401": {
            "description": "Authentication required or invalid",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "403": {
            "description": "Authenticated but not permitted",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "404": {
            "description": "Requested resource not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "409": {
            "description": "Resource state conflict",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "422": {
            "description": "Field validation failed",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "429": {
            "description": "Rate or attempt limit exceeded",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "503": {
            "description": "External delivery service unavailable",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          }
        },
        "security": [{ "HTTPBearer": [] }]
      }
    },
    "/api/v1/home": {
      "get": {
        "tags": ["Home and Content"],
        "summary": "Get the authenticated beneficiary home dashboard",
        "operationId": "get_home_api_v1_home_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HomeDashboardResponse"
                }
              }
            }
          },
          "400": {
            "description": "Business rule or invalid request",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "401": {
            "description": "Authentication required or invalid",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "403": {
            "description": "Authenticated but not permitted",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "404": {
            "description": "Requested resource not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "409": {
            "description": "Resource state conflict",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "422": {
            "description": "Field validation failed",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "429": {
            "description": "Rate or attempt limit exceeded",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "503": {
            "description": "External delivery service unavailable",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          }
        },
        "security": [{ "HTTPBearer": [] }]
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiError": {
        "properties": {
          "message": { "type": "string", "title": "Message" },
          "code": { "type": "string", "title": "Code" },
          "fieldErrors": {
            "anyOf": [
              {
                "additionalProperties": { "type": "string" },
                "type": "object"
              },
              { "type": "null" }
            ],
            "title": "Fielderrors"
          }
        },
        "type": "object",
        "required": ["message", "code"],
        "title": "ApiError"
      },
...
      "BeneficiaryEventResponse": {
        "properties": {
          "id": { "type": "string", "title": "Id" },
          "name": { "type": "string", "title": "Name" },
          "dateLabel": { "type": "string", "title": "Datelabel" },
          "timeLabel": { "type": "string", "title": "Timelabel" },
          "location": { "type": "string", "title": "Location" },
          "icon": { "type": "string", "title": "Icon" },
          "startsAt": {
            "type": "string",
            "format": "date-time",
            "title": "Startsat"
          },
          "endsAt": {
            "type": "string",
            "format": "date-time",
            "title": "Endsat"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "dateLabel",
          "timeLabel",
          "location",
          "icon",
          "startsAt",
          "endsAt"
        ],
        "title": "BeneficiaryEventResponse"
      },
      "BeneficiaryProfileResponse": {
        "properties": {
          "id": { "type": "string", "title": "Id" },
          "name": { "type": "string", "title": "Name" },
          "birthDate": { "type": "string", "title": "Birthdate" },
          "stage": { "type": "string", "title": "Stage" },
          "address": { "type": "string", "title": "Address" },
          "phone": { "type": "string", "title": "Phone" },
          "whatsapp": { "type": "string", "title": "Whatsapp" },
          "school": { "type": "string", "title": "School" },
          "classSaintName": { "type": "string", "title": "Classsaintname" },
          "confessionFather": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Confessionfather"
          },
          "talents": {
            "items": { "type": "string" },
            "type": "array",
            "title": "Talents"
          },
          "email": { "type": "string", "format": "email", "title": "Email" },
          "points": { "type": "integer", "title": "Points" },
          "attendanceQrValue": {
            "type": "string",
            "title": "Attendanceqrvalue"
          },
          "avatarUri": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Avataruri"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "birthDate",
          "stage",
          "address",
          "phone",
          "whatsapp",
          "school",
          "classSaintName",
          "talents",
          "email",
          "points",
          "attendanceQrValue"
        ],
        "title": "BeneficiaryProfileResponse"
      },
...
      "DailyReadingResponse": {
        "properties": {
          "id": { "type": "string", "title": "Id" },
          "date": { "type": "string", "title": "Date" },
          "reference": { "type": "string", "title": "Reference" },
          "title": { "type": "string", "title": "Title" },
          "content": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Content"
          }
        },
        "type": "object",
        "required": ["id", "date", "reference", "title"],
        "title": "DailyReadingResponse"
      },
...
      "HomeDashboardResponse": {
        "properties": {
          "profile": {
            "$ref": "#/components/schemas/BeneficiaryProfileResponse"
          },
          "upcomingEvents": {
            "items": {
              "$ref": "#/components/schemas/BeneficiaryEventResponse"
            },
            "type": "array",
            "title": "Upcomingevents"
          },
          "dailyReading": {
            "anyOf": [
              { "$ref": "#/components/schemas/DailyReadingResponse" },
              { "type": "null" }
            ]
          },
          "questionCategories": {
            "items": {
              "$ref": "#/components/schemas/QuestionCategoryResponse"
            },
            "type": "array",
            "title": "Questioncategories"
          },
          "knowMeQuestions": {
            "items": { "$ref": "#/components/schemas/KnowMeQuestionResponse" },
            "type": "array",
            "title": "Knowmequestions"
          }
        },
        "type": "object",
        "required": [
          "profile",
          "upcomingEvents",
          "dailyReading",
          "questionCategories",
          "knowMeQuestions"
        ],
        "title": "HomeDashboardResponse"
      },
...
      "KnowMeQuestionResponse": {
        "properties": {
          "id": { "type": "string", "title": "Id" },
          "label": { "type": "string", "title": "Label" },
          "askedAt": { "type": "string", "title": "Askedat" },
          "isToday": { "type": "boolean", "title": "Istoday" },
          "points": { "type": "integer", "title": "Points" },
          "answered": {
            "type": "boolean",
            "title": "Answered",
            "default": false
          }
        },
        "type": "object",
        "required": ["id", "label", "askedAt", "isToday", "points"],
        "title": "KnowMeQuestionResponse"
      },
...
      "QuestionCategoryResponse": {
        "properties": {
          "id": { "type": "string", "title": "Id" },
          "title": { "type": "string", "title": "Title" },
          "icon": { "type": "string", "title": "Icon" },
          "color": { "type": "string", "title": "Color" }
        },
        "type": "object",
        "required": ["id", "title", "icon", "color"],
        "title": "QuestionCategoryResponse"
      },
...
    },
    "securitySchemes": { "HTTPBearer": { "type": "http", "scheme": "bearer" } }
  }
}